Revert "Make SHA256_update and SHA512_update if the passed in length is greater than UINT32_MAX"
Needed to unbreak the tegra build.

This reverts commit a01a62bf85e01e1687ea5132286628b2a760f6d8.

R=gauravsh@chromium.org
BUG=
TEST=

Review URL: http://codereview.chromium.org/6731068

+++ firmware/lib/cryptolib/include/sha.h
@@ -58,11 +58,11 @@ void SHA1_update(SHA1_CTX* ctx, const uint8_t* data, uint64_t len);
 uint8_t* SHA1_final(SHA1_CTX* ctx);
 
 void SHA256_init(SHA256_CTX* ctx);
-void SHA256_update(SHA256_CTX* ctx, const uint8_t* data, uint32_t len);
+void SHA256_update(SHA256_CTX* ctx, const uint8_t* data, uint64_t len);
 uint8_t* SHA256_final(SHA256_CTX* ctx);
 
 void SHA512_init(SHA512_CTX* ctx);
-void SHA512_update(SHA512_CTX* ctx, const uint8_t* data, uint32_t len);
+void SHA512_update(SHA512_CTX* ctx, const uint8_t* data, uint64_t len);
 uint8_t* SHA512_final(SHA512_CTX* ctx);
 
 /* Convenience function for SHA-1.  Computes hash on [data] of length [len].
